feat(race): add Race, mod Game 경주 진행 분리
